pyproject.toml
[tool.poetry]
name = "alkuaanet"
version = "0.1.0"
description = ""
authors = ["Simo Hellgrén <hellgrensimo@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^1.25.2"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
flake8 = "^6.1.0"
boto3 = "^1.28.66"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
click = "^8.1.8"


[tool.poetry.group.telegram.dependencies]
telepot = "^12.7"
requests = "^2"


[tool.poetry.group.graph.dependencies]
mangum = "^0.19"
strawberry-graphql = {extras = ["asgi"], version = "^0"}
pydantic = "^2.10.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
